requirements.txt
alembic==1.7.7
aniso8601==9.0.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
cachelib==0.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
DBUtils==1.4
Deprecated==1.2.13
elastic-transport==8.1.0
elasticsearch==8.1.1
Flask==2.0.3
Flask-APScheduler==1.12.3
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Limiter==2.2.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
gevent-websocket==0.10.1
google-auth==2.6.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
kubernetes==23.3.0
limits==2.4.0
Mako==1.2.0
MarkupSafe==2.1.0
mysql-connector==2.2.9
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.32
tornado==6.1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
